requirements.txt
pip~=23.2.1
wheel~=0.41.2
setuptools~=68.0.0
moviepy
pandas
pillow
numpy
gTTS